Add print-dupkeybinds
The spacing in the help Makefile recipe has been padded to account for
the filename size of print-dupkeybinds (print duplicate keybindings).

The SCHEMAS value was derived from running the following command
'gsettings list-schemas | grep keybindings'. I felt this was a good
enough list to start with for schemas that were specific to keybindings.