Add GNOME Display Manager (gdm) configurations
The spacing in the help Makefile recipe has been padded to account for
the install-gdm-configs Makefile target size.

I derived installing these configurations mostly from the ArchWiki (see
https://wiki.archlinux.org/title/GDM) primarily because GNOME's
documentation on gdm is not up-to-date. This was pointed out when
reading https://gitlab.gnome.org/GNOME/gdm/-/issues/560. There is an
option to install the configurations by swapping to the 'gdm' user and
running gsettings commands, but I opted not todo this because this
seemed unusual for configuring system-level software. The keyfile
filename '00-desktop-configs' is used because I only anticipate there
being only one keyfile used to configure gdm.