Install Visual Studio Code workspaces
The 'personal' workspace will be for repositories I maintain and am
actively working on.

The spacing in the help Makefile recipe has been padded to account for
the install-vscode-workspaces Makefile target size.

A script was not created for installing the Visual Studio Code
workspaces because installing them is straightforward (e.g. no need to
check for external executables).