Add script to set Nautilus file manager icons
The custom icons are set from /usr/share/icons primarily because that is
where the papirus-icon-theme deb package installs to (instead of
~/.local/share/icons). Furthermore, while a 64x64 size is used for each
icon, this shouldn't matter because SVGs can be scaled to any size
necessary.

It was noticed that the gio executable does not error out when the
icon_path does not exist, but does error out when location_path does not
exist. Hence, why the error message displayed is only related to the
latter. The error message was inspired by the message the ls executable
emits when given a nonexistent file or directory.