Read in local variant of the gsettings.txt file
This was done for the apply-gsettings script to override keys based on
the development machine that I'd be working on. The sync-gsettings
script was not modified to account for these changes because it is
intended that only a few to some keys will be written to the local
gsettings.txt file. Hence, there isn't much value in syncing the
gsettings to this other file.