Add sshd configurations to solely use pubkey auth
The keywords' order in the sshd_pubkey_auth.conf file follows their
order within the default sshd_config(5) (/etc/ssh/sshd_config) file. The
AuthorizedKeysFile's value uses a single space vs the several in the
default sshd_config file because I find the additional spacing to be
unnecessary and not something enforced when reading this keyword's
description in sshd_config(5).

The spacing in the help Makefile recipe has been padded to account for
the install-sshd-pubkey-auth-conf Makefile target size.