Replace the jq processor with the yq processor
This was done because I'd like to show more support to yq by using it
over jq. I ended up using yq recently due to needing to output a JSON
blob to a file without using the shell todo this (e.g. redirection or
pipe to an executable that can write to a file). Unfortunately, jq
doesn’t offer an option like this, much like the output option for curl.
Also, I went down the rabbit hole to find that the maintainers of jq
would be resistant to the addition of this feature.