Add the pre-commit framework to the repository
Despite being resistant to adding any language (e.g. Python, Ruby)
package dependencies to the repository, I am fine with relaxing this
stance for pre-commit because it is vital for maintaining most of my
repositories. I did explore using the pre-commit package from my
distribution's repositories; however, the version (2.17.0) it has runs
with an error. A remedy suggested for the error is to just upgrade the
pre-commit version (as mentioned here,
https://github.com/psf/black/issues/4065#issuecomment-1821277094).