Set VSCode's maximum number of watched files
This is being done so VSCode/VSCodium doesn't likely complain about
being able to watch file changes in a repository. The fix for this can
be found here under the '"Visual Studio Code is unable to watch for file
changes in this large workspace" (error ENOSPC)' section,
https://code.visualstudio.com/docs/setup/linux#_common-questions.

A '99-' numerical prefix is used in the destination for the
sysctl.conf(5) file because in the event a package chooses to add
a sysctl.conf(5) file, it's unlikely any parameters set by a package
will override the parameters I set. This is because I would expect mine
to be read in after any packages' sysctl.conf(5) file.