Set VSCode's maximum number of watched files
This is being done so VSCode/VSCodium doesn't likely complain about
being able to watch file changes in a repository. The fix for this can
be found here, https://go.microsoft.com/fwlink/?linkid=867693.

A '99-' numerical prefix is used in the destination filename for the
sysctl.conf(5) file because in the event a package chooses to add
a sysctl.conf(5) file. It's unlikely any parameters set by a package
will override the parameters I set, as I would expect mine to be read in
after a package's file.