Add a hosts(5) file to be copied into /etc/hosts
This file is being added primarily because some of my development
machines were unable to resolve their hostname when connected to a
network where the DNS server advertised was not the same server running
DHCP. Machines that obtain an IP address via a DHCP server will likely
have their hostname registered as a DNS record if a DNS daemon also runs
on said DHCP server.

The IP address of 127.0.1.1 is used for my development machines
hostnames because per hosts(5), this is often used.