Add Nautilus bookmarks for VSCode exts and icons
The original intention was to have bookmarks for the VSCode vscode-icons
extension and my custom VSCode icons directory. However, with the
former, each extension directory has a suffix containing the version of
the extension. Meaning, it isn't trivial to hardcode a bookmark for any
extension. Hence, I chose to just bookmark the extensions (exts)
directory.