Move all desktop configuration files into a subdir
This was done because I've felt that over time the top-level of this Git
repository was starting to get crowded with my desktop configuration
files and those related to the tooling (e.g. shellcheck) of this
repository.

The bookmarks shell template for the Nautilus file manager was moved out
of its directory because I don't currently foresee any additional files
being created to configure Nautilus.