Add config files for local DNS resolving
This was done because, as related to fbdb486, I learned that Systemd has
Name Service Switch (NSS) plug-in modules for resolving a machine's
local, configured hostname. These would include the nss-resolve and
nss-myhostname modules. However, only the latter made it into the
nsswitch.conf file, whose source order in the hosts database is based on
the recommendation from nss-myhostname(8). Its addition is despite there
being conversation about intent to remove the module (see
https://github.com/systemd/systemd/issues/1605#issuecomment-149306837).
The nss-myhostname module mimics the behavior of synthetic records
performed by the systemd-resolved service.

The nss-resolve module was not integrated because, despite
nss-resolve(8) giving the impression that the module will NOT use DNS to
resolve a hostname (via the service, when placed in the source order
recommended by nss-resolve(8)), that isn't the case. In this context, I
am defining a hostname to be a name with no dots (excludes a name
appended with the root domain). Instead, the service will still try to
resolve any hostname via DNS (with search domain logic) before possibly
reaching for a synthetic record. This assertion is primarily supported
by my testing and the conversation at
https://github.com/systemd/systemd/issues/7160#issuecomment-338735051.
A scenario that I have experienced with this is if a per-interface DNS
server is down, this has caused sudo to hang because it, by default on
Pop!_OS 22.04, resolves the local hostname via DNS (using glibc's
getaddrinfo, which eventually grabs the nss-resolve module). This
behavior of the nss-resolve module with the systemd-resolved service
could be due to the version of Systemd my machines run (249). It could
also explain why Pop!_OS 22.04 doesn't install this module by default.

Within the hosts database, I explicitly chose to use the mdns4_minimal
service specification over mdns4, as I only need to resolve .local
domains and not include other domains for mDNS. This also presumes I
will be using Avahi for mDNS resolving and not the systemd-resolved
service.

The rest of the service specifications following each database I have
reviewed and am fine with for now. The systemd plug-in module for the
passwd and groups databases can be useful for root and nobody groups
always existing even if respective files don't.