Replace cron jobs with systemd timers and services
I did this because my homelab hosts have used these systemd units, even
when it would be acceptable to utilize cron jobs. That, and I don't have
a strong personal preference for using the cron daemon.

The systemd.unit(8) filenames that contain a suffix that ends with '@'
before the type suffix are also known as template units. Essentially,
they are templates to instantiate other units. For example, since I
intend on having my previous cron jobs execute every hour, I created an
hourly template timer unit that would be used for each cron job timer.
Eliminating the need to create a timer unit file for each cron job. This
timer unit uses the default.target unit as its value to the WantedBy
option because neither the multi-user.target nor graphical.target units
are defined in a user instance of systemd (see UNITS MANAGED BY THE
USER SERVICE MANAGER section of systemd.special(7)).

A template unit file was created to send mail/email with the latest
systemd(1) journal logs for a service because this is something not
natively supported by systemd. Unfortunately, this is a known
limitation: https://wiki.archlinux.org/title/Systemd/Timers#Caveats.

A subdirectory for systemd was created under the libexec directory so I
can clearly indicate that the sendmail script is to be used by systemd
units. The alternative was prefixing the script's filename with
'systemd-', which I wasn't a fan of. To add, many others also use this
approach (e.g. coreutils, dpkg, docker).