Remove the org.gnome.metacity.keybindings schema
This was done because none of my development machines (which run the
Pop!_OS 22.04 operating system) as of today have this schema installed.
A previous installation of Pop!_OS (whose version was before 22.04) most
likely installed the Metacity window manager, which would have then
installed the aforementioned schema. Metacity has been superseded by the
Mutter window manager as of GNOME 3 and is the default window manager
for the GNOME desktop environment today.